+++ packages/tgweb/lib/tgweb/generate_html.mjs
@@ -1,10 +1,12 @@
 import getType from "./get_type.mjs"
-import generationFuncs from "./generation_funcs.mjs"
+import { renderPage } from "./render_page.mjs"
+import { renderArticle } from "./render_article.mjs"
 
 const generateHTML = (path, siteData) => {
   const type = getType(path)
 
-  if (typeof generationFuncs[type] === "function") return generationFuncs[type](path, siteData)
+  if (type === "page") return renderPage(path, siteData)
+  else if (type === "article") return renderArticle(path, siteData)
 }
 
 export default generateHTML