+++ packages/tgweb/lib/tgweb/get_title.mjs
@@ -1,0 +1,23 @@
+const getTitle = (template, element) => {
+  if (template.frontMatter["title"]) return template.frontMatter["title"]
+
+  const h1 = element.querySelector("h1")
+  if (h1) return h1.textContent
+
+  const h2 = element.querySelector("h2")
+  if (h2) return h2.textContent
+
+  const h3 = element.querySelector("h3")
+  if (h3) return h3.textContent
+
+  const h4 = element.querySelector("h4")
+  if (h4) return h4.textContent
+
+  const h5 = element.querySelector("h5")
+  if (h5) return h5.textContent
+
+  const h6 = element.querySelector("h6")
+  if (h6) return h6.textContent
+}
+
+export { getTitle }