+++ packages/tgweb/lib/tgweb/update.mjs
@@ -1,16 +1,28 @@
-import * as PATH from "path"
-import fs from "fs"
 import { updateSiteData } from "./update_site_data.mjs"
 import getType from "./get_type.mjs"
-import generateHTML from "./generate_html.mjs"
+import { updateHTML } from "./update_html.mjs"
+import { dbg, pp } from "./debugging.mjs"
+
+// Prevent warnings when functions dbg and pp are not used.
+if (dbg === undefined) { dbg() }
+if (pp === undefined) { pp() }
 
 const update = (path, siteData) => {
   updateSiteData(siteData, path)
 
   const type = getType(path)
 
-  if (type === "page" || type === "article") {
+  if (type === "page") {
     updateHTML(path, siteData)
+  }
+  else if (type === "article") {
+    updateHTML(path, siteData)
+
+    const name = path.replace(/^src\//, "").replace(/\.html$/, "")
+
+    siteData.pages
+      .filter(page => page.dependencies.includes(name))
+      .forEach(page => updateHTML("src/pages/" + page.path, siteData))
   }
   else if (type == "site.yml") {
     siteData.pages.forEach(page => updateHTML("src/pages/" + page.path, siteData))
@@ -29,16 +41,4 @@ const update = (path, siteData) => {
   }
 }
 
-const updateHTML = (path, siteData) => {
-  const html = generateHTML(path, siteData)
-
-  if (html !== undefined) {
-    const targetPath = path.replace(/^src\//, "dist/").replace(/^dist\/pages\//, "dist/")
-    const targetDir = PATH.dirname(targetPath)
-
-    if (!fs.existsSync(targetDir)) fs.mkdirSync(targetDir, { recursive: true })
-    fs.writeFileSync(targetPath, html)
-  }
-}
-
-export default update
+export { update }